+++ app/types/store.js
@@ -1,3 +1,5 @@
+// @flow
+
 import type {
   Store as ReduxStore,
   Dispatch as ReduxDispatch
