+++ app/types/user.js
@@ -1,3 +1,5 @@
+// @flow
+
 import type { OfflineMeta } from 'types/offline';
 
 import {
@@ -9,16 +11,27 @@ import {
   LOGOUT_REQUEST
 } from 'redux-modules/user';
 
+export type UserState = {
+  data: Object,
+  loggedIn: boolean,
+  token: ?string,
+  oAuthToken: ?string,
+  socialNetwork: ?string,
+  logSuccess: boolean,
+  synced: boolean,
+  syncing: boolean
+}
+
 export type UserAction = GetUserRequest | GetUserCommit | GetUserRollback | SetLoginAuth | SetLoginStatus | LogoutRequest;
 
 type User = {};
 
-type GetUserRequest = { type: GET_USER_REQUEST, meta: OfflineMeta };
-type GetUserCommit = { type: GET_USER_COMMIT, payload: User };
-type GetUserRollback = { type: GET_USER_ROLLBACK };
+type GetUserRequest = { type: typeof GET_USER_REQUEST, meta: OfflineMeta };
+type GetUserCommit = { type: typeof GET_USER_COMMIT, payload: User };
+type GetUserRollback = { type: typeof GET_USER_ROLLBACK };
 
 type SetLoginAuth = {
-  type: SET_LOGIN_AUTH,
+  type: typeof SET_LOGIN_AUTH,
   payload: {
     socialNetwork: string,
     loggedIn: boolean,
@@ -28,10 +41,10 @@ type SetLoginAuth = {
 }
 
 type SetLoginStatus = {
-  type: SET_LOGIN_STATUS,
+  type: typeof SET_LOGIN_STATUS,
   payload: boolean
 }
 
 type LogoutRequest = {
-  type: LOGOUT_REQUEST
+  type: typeof LOGOUT_REQUEST
 }
